Use square brackets for environment variable placeholders
The first time I did this I included the quotes, and nothing worked. It awasn't until I tried debugging it that I realised that the quotes were being included as part of the value.

By using square brackets instead of single quotes then hopefully it will be clearer that the delimiter is not part of the expected final output.